refactor: move it to a separate group
