+++ package.json
@@ -34,9 +34,9 @@
     "use-onclickoutside": ">= 0.3.1"
   },
   "devDependencies": {
-    "@testing-library/dom": "7.20.0",
-    "@testing-library/jest-dom": "5.11.0",
-    "@testing-library/react": "10.4.5",
+    "@testing-library/dom": "7.21.0",
+    "@testing-library/jest-dom": "5.11.1",
+    "@testing-library/react": "10.4.7",
     "@testing-library/user-event": "12.0.11",
     "@types/jest": "26.0.4",
     "@types/node": "12.12.38",
@@ -45,8 +45,8 @@
     "@types/react-portal": "4.0.2",
     "@types/react-responsive": "8.0.2",
     "@types/styled-components": "5.1.1",
-    "@typescript-eslint/eslint-plugin": "3.6.0",
-    "@typescript-eslint/parser": "3.6.0",
+    "@typescript-eslint/eslint-plugin": "3.6.1",
+    "@typescript-eslint/parser": "3.6.1",
     "babel-eslint": "10.1.0",
     "cross-env": "7.0.2",
     "eslint": "7.4.0",
@@ -58,7 +58,7 @@
     "eslint-plugin-prettier": "3.1.4",
     "eslint-plugin-promise": "4.2.1",
     "eslint-plugin-react": "7.20.3",
-    "eslint-plugin-react-hooks": "4.0.7",
+    "eslint-plugin-react-hooks": "4.0.8",
     "eslint-plugin-standard": "4.0.1",
     "microbundle-crl": "0.13.11",
     "npm-run-all": "4.1.5",
