+++ src/components/checkbox/index.tsx
@@ -3,3 +3,4 @@ export * from './use-checkbox'
 export * from './radio'
 export * from './radiogroup'
 export * from './toggle'
+export * from './use-radiogroup'
