+++ src/components/checkbox/radiogroup.tsx
@@ -1,43 +1,7 @@
-import React, { createContext, useState, useMemo } from 'react'
+import React from 'react'
 import { Flex } from '../flex'
-import nextId from 'react-id-generator'
-
-interface RadioGroupState {
-  name: string
-
-  setName: Function
-
-  value: string
-
-  setValue: Function
-}
-
-export const RadioGroupContext = createContext<RadioGroupState>({
-  name: '',
-  setName: () => {},
-  value: '',
-  setValue: () => {}
-})
-
-export function useRadioGroup(
-  pValue?: string,
-  pName?: string
-): RadioGroupState {
-  const actualName =
-    pName !== undefined && pName !== null ? pName : nextId('rg-name-')
-  const actualValue = pValue !== undefined && pValue !== null ? pValue : ''
-  const [name, setName] = useState(actualName)
-  const [value, setValue] = useState(actualValue)
-  return useMemo(
-    () => ({
-      name,
-      setName,
-      value,
-      setValue
-    }),
-    [name, setName, value, setValue]
-  )
-}
+// eslint-disable-next-line no-unused-vars
+import { RadioGroupContext, RadioGroupState } from './use-radiogroup'
 
 export interface RadioGroupProps {
   children: React.ReactNode
