+++ src/index.tsx
@@ -14,6 +14,7 @@ export * from './components/text'
 export * from './components/tooltip'
 export * from './contexts/mq'
 export * from './mixins/disclosure'
+export * from './mixins/display'
 export * from './mixins/color'
 export * from './mixins/size'
 export * from './theming'
